node-installer: move flavours to node-installer
Since the main consumer of the flavours is the node-installer, the package now lives here. It also has the benefit of not needing to rebuild the node-installer whenever Contrast changes, but the other way around (which should be the case way less often).
