+++ e2e/internal/contrasttest/contrasttest.go
@@ -20,9 +20,9 @@ import (
 
 	"github.com/edgelesssys/contrast/cli/cmd"
 	"github.com/edgelesssys/contrast/e2e/internal/kubeclient"
-	"github.com/edgelesssys/contrast/internal/flavours"
 	"github.com/edgelesssys/contrast/internal/kubeapi"
 	"github.com/edgelesssys/contrast/internal/kuberesource"
+	"github.com/edgelesssys/contrast/node-installer/flavours"
 	ksync "github.com/katexochen/sync/api/client"
 	"github.com/stretchr/testify/require"
 )
