+++ node-installer/flavours/flavours.go
