+++ packages/by-name/contrast/package.nix
@@ -53,17 +53,16 @@ buildGoModule rec {
         (path.append root "go.mod")
         (path.append root "go.sum")
         (path.append root "cli/cmd/assets/image-replacements.txt")
+        (path.append root "node-installer")
         (fileset.difference
           (fileset.fileFilter (file: hasSuffix ".go" file.name) root)
-          (fileset.unions [
-            (path.append root "service-mesh")
-            (path.append root "node-installer")
-          ]))
+          (path.append root "service-mesh")
+        )
       ];
     };
 
   proxyVendor = true;
-  vendorHash = "sha256-9RtHcArCyJqvC3VfNOyuyeN9+KyVu+wWKpSxXoCj/3o=";
+  vendorHash = "sha256-KJwQuK7SAY8qSmQhTdSKgE9Z9f3/e9I/J0w6j9YURwc=";
 
   subPackages = packageOutputs ++ [ "internal/kuberesource/resourcegen" ];
